Core: Ports, initiated cleanup / refactoring, moved color_code into data. Each data type such as volume, image, should now have a static member COLOR_CODE that is used by the ports.  closes inviwo/inviwo-dev#406

+++ modules/basegl/processors/volumeraycaster.cpp
@@ -100,7 +100,8 @@ void VolumeRaycaster::onVolumeChange() {
 
         std::vector<OptionPropertyIntOption> channelOptions;
         for (size_t i = 0; i < channels; i++) {
-            channelOptions.emplace_back("Channel " + toString(i), "Channel " + toString(i), i);
+            channelOptions.emplace_back("Channel " + toString(i), "Channel " + toString(i),
+                                        static_cast<int>(i));
         }
         channel_.replaceOptions(channelOptions);
         channel_.setCurrentStateAsDefault();
